# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"DefaultResponse_string_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"CurrencyProviderResponse": {
				"properties": {
					"error": {
						"type": "number",
						"format": "double"
					},
					"error_message": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"DefaultResponse_CurrencyProviderResponse_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/CurrencyProviderResponse"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "rest-api-redis",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/api_currency": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DefaultResponse_string_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"data": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DefaultResponse_string_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "string"
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"CurrencyExchange"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api_currency/getCurrency": {
			"get": {
				"operationId": "GetCurrency",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DefaultResponse_CurrencyProviderResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"CurrencyExchange"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "from",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "to",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "amount",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	],
	"tags": [
		{
			"name": "CurrencyExchange",
			"description": "Currency Exchange Service",
			"externalDocs": {
				"description": "Find out more here on official currency exchange docs.",
				"url": "https://www.amdoren.com/currency-api/"
			}
		}
	]
}